refactor(votingSystem): rename generateRSAKeys -> generateKeyPair
